Refactoring auth and metric interceptors
